Move videoembed tests to their own class
